@@ -5,49 +5,71 @@
   "angular-cli": {},
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
+    "start": "node server.js",
     "lint": "tslint \"src/**/*.ts\" --project src/tsconfig.json --type-check && tslint \"e2e/**/*.ts\" --project e2e/tsconfig.json --type-check",
     "test": "ng test",
     "pree2e": "webdriver-manager update --standalone false --gecko false",
-    "e2e": "protractor"
+    "e2e": "protractor",
+    "postinstall": "ng build --aot -prod"
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "4.0.0",
-    "@angular/compiler": "4.0.0",
-    "@angular/core": "4.0.0",
-    "@angular/forms": "4.0.0",
-    "@angular/http": "4.0.0",
-    "@angular/platform-browser": "4.0.0",
-    "@angular/platform-browser-dynamic": "4.0.0",
-    "@angular/router": "4.0.0",
+    "@angular/animations": "6.0.0",
+    "@angular/cli": "6.0.0",
+    "@angular/common": "6.0.0",
+    "@angular/compiler": "6.0.0",
+    "@angular/compiler-cli": "6.0.0",
+    "@angular/core": "6.0.0",
+    "@angular/forms": "6.0.0",
+    "@angular/platform-browser": "6.0.0",
+    "@angular/platform-browser-dynamic": "6.0.0",
+    "@angular/platform-server": "6.0.0",
+    "@angular/router": "6.0.0",
     "admin-lte": "2.3.11",
-    "core-js": "2.4.1",
+    "ajv": "6.0.0",
+    "core-js": "2.5.4",
+    "express": "^4.17.1",
     "font-awesome": "4.7.0",
     "intl": "1.2.5",
     "jquery": "3.1.1",
-    "reflect-metadata": "0.1.9",
-    "rxjs": "5.1.0",
+    "path": "^0.12.7",
+    "reflect-metadata": "0.1.10",
+    "rxjs": "6.0.0",
+    "rxjs-compat": "6.0.0",
     "ts-helpers": "1.1.2",
-    "zone.js": "0.8.4"
+    "web-animations-js": "2.2.5",
+    "zone.js": "0.8.26",
+    "typescript": "2.7.2"
   },
   "devDependencies": {
-    "@angular/cli": "1.2.7",
-    "@angular/compiler-cli": "4.0.0",
-    "@types/jasmine": "2.5.38",
-    "@types/node": "7.0.5",
-    "codelyzer": "2.0.0",
-    "jasmine-core": "2.5.2",
-    "jasmine-spec-reporter": "3.2.0",
-    "karma": "1.4.1",
-    "karma-chrome-launcher": "2.0.0",
-    "karma-cli": "1.0.1",
-    "karma-jasmine": "1.1.0",
-    "karma-remap-istanbul": "0.2.2",
-    "protractor": "5.1.0",
-    "ts-node": "2.1.0",
-    "tslint": "4.5.0",
-    "typescript": "2.2.0",
-    "webdriver-manager": "10.2.5"
+    "@angular-devkit/build-angular": "~0.6.0",
+    "@angular/cli": "6.0.0",
+    "@angular/compiler-cli": "6.0.0",
+    "@angular/language-service": "6.0.0",
+    "@types/express": "4.0.37",
+    "@types/jasmine": "2.8.6",
+    "@types/jasminewd2": "2.0.3",
+    "@types/jsonwebtoken": "7.2.3",
+    "@types/node": "8.9.4",
+    "codelyzer": "4.2.1",
+    "enhanced-resolve": "^3.3.0",
+    "jasmine-core": "2.99.1",
+    "jasmine-spec-reporter": "4.2.1",
+    "json-server": "0.12.0",
+    "jsonwebtoken": "7.4.1",
+    "karma": "1.7.1",
+    "karma-chrome-launcher": "2.2.0",
+    "karma-coverage-istanbul-reporter": "1.4.2",
+    "karma-jasmine": "1.1.1",
+    "karma-jasmine-html-reporter": "0.2.2",
+    "protractor": "5.3.0",
+    "ts-node": "5.0.1",
+    "tslint": "5.9.1",
+    "typescript": "2.7.2",
+    "webdriver-manager": "12.0.6"
+  },
+  "engines": {
+    "node": "10.19.0",
+    "npm": "6.13.4"
   }
 }
